Add node_ancestors method (#1169)
I've been thinking about ways to get rid of unbounded `while` loops in
the Pylint codebase. A common use is to loop over a node's ancestors.
The `node_ancestors` method centralizes this logic in one place so
that those `while` loops can be rewritten as `for`.

A few uses are made of this new method. It only comes up a few places
in Astroid, but there are many more in Pylint.