Add recognition of previous partial args/kwargs when chaining
Repeat use of functools.partial on the same function would only
keep the args & kwargs locked in by the last call to partial.
By checking if the wrapped function is itself a partial, and
including its filled args/kwargs if it is, we can keep track
of all locked args/kwargs across any number of nested partial
calls.