Move ``_filter_stmts`` into its own file (#1303)
We need access to all types of nodes in _filter_stmts. Not only those in node_classes
(the file in which it was defined).

This change shows that by moving _filter_stmts into its own file we can import the nodes
module and access all of them. This will allow to fix some issue with it. I have tried to keep
the change as minimal as possible and only changed how we refer to nodes.