Set mypy to strict, allow untyped decorators
These are the settings I've been using for adding types. Obviously
we're far from getting it to pass with strict settings, but strict
makes it easier to see what is unannotated.

For me at least, adding types to decorators is totally unfeasible
right now, so those should be allowed for the time being.