brain_attrs: Support annotation-only members (#2515)
Similar to dataclasses, the following class which uses instance
variable annotations is valid:

```py
@attrs.define
class AttrsCls:
    x: int
AttrsCls(1).x
```

However, before this commit astroid failed to transform the class
attribute into an instance attribute and this led to `no-member` errors
in pylint.

Only the new `attrs` API supports this form out-of-the-box, so just
address the common case.

Closes #2514